local_mcp_server/requirements.txt
fastapi
uvicorn
python-dotenv
pyodbc
pydantic
langchain
langchain-openai